refactor(dotcom): split admin panel into sidebar sections
